update e3.bash and maintain_e3.bash from E3+ branch

+++ e3.bash
@@ -19,13 +19,15 @@
 #
 #   author  : Jeong Han Lee
 #   email   : jeonghan.lee@gmail.com
-#   date    : Monday, February 12 15:39:11 CET 2018
-#   version : 0.0.8
+#   date    : Wednesday, April 18 09:36:32 CEST 2018
+#   version : 0.2.3
+
 
 
 declare -gr SC_SCRIPT="$(realpath "$0")"
 declare -gr SC_SCRIPTNAME=${0##*/}
-declare -gr SC_TOP="$(dirname "$SC_SCRIPT")"
+declare -gr SC_TOP="${SC_SCRIPT%/*}"
+
 
 function pushd { builtin pushd "$@" > /dev/null; }
 function popd  { builtin popd  "$@" > /dev/null; }
@@ -35,16 +37,14 @@ GIT_URL="https://github.com/icshwi"
 GIT_CMD="git clone"
 BOOL_GIT_CLONE="TRUE"
 
-declare -g  env="e3-env"
-declare -ga require_list=("e3-base" "e3-require")
+
+declare -ga base_list=("e3-base")
+declare -ga require_list=("e3-require")
 declare -ga module_list=()
-declare -ga db_module_list=()
+declare -ga m_list=()
 
-db_module_list+=("e3-iocStats");
-db_module_list+=("e3-mrfioc2");
-db_module_lits+=("e3-ipmiComm");
 
-function die()
+function die
 {
     error=${1:-1}
     ## exits with 1 if error number not given
@@ -54,65 +54,32 @@ function die()
     exit "$error"
 }
 
-
-function env_reset {
-    
-    PATH=/usr/local/bin:/usr/bin:/bin
-    unset EPICS_MODULES_PATH
-    unset EPICS_ENV_PATH
-    unset EPICS_BASE
-    unset EPICS_BASES_PATH
-    unset EPICS_HOST_ARCH
-    unset PYTHONPATH
-    
-    unset EPICS_CA_AUTO_ADDR_LIST
-}
+EXIST=1
+NON_EXIST=0
 
 
-
-
-function yes_or_no_to_go() {
-
-    printf "\n";
-    printf  ">>>> $1\n";
-    read -p ">>>> Do you want to continue (y/n)? " answer
-    case ${answer:0:1} in
-	y|Y )
-	    printf ">>>> We are installing modules ...... \n";
-	    ;;
-	* )
-            printf "Stop here.\n";
-	    exit;
-    ;;
-    esac
-
-}
-
-function install_db
+function checkIfDir
 {
-    local rep;
-    for rep in  ${db_module_list[@]}; do
-	pushd ${rep}
-	make db
-	make install
-	popd
-    done
+    
+    local dir=$1
+    local result=""
+    if [ ! -d "$dir" ]; then
+	result=$NON_EXIST
+	# doesn't exist
+    else
+	result=$EXIST
+	# exist
+    fi
+    echo "${result}"
+};
+
+
+function checkout_e3_tags
+{
+    local tags=$1;
+    git checkout $tags;
 }
 
-function help
-{
-    echo "">&2
-    echo " Usage: $0 <arg>  ">&2 
-    echo ""
-    echo "          <arg>    : info">&2 
-    echo ""
-    echo "           all     : Setup and Build ALL">&2
-    echo "           base    : Setup and Build Base and Require">&2
-    echo "           modules : Setup and Build all modules">&2
-    echo "           env     : Print all modules list">&2
-    echo "">&2
-    echo "">&2 	
-}
 
 function git_clone
 {
@@ -124,8 +91,9 @@ function git_clone
 
 function get_module_list
 {
-    local i;
-    let i=0
+    local empty_string="";
+    declare -a entry=();
+ 
     while IFS= read -r line_data; do
 	if [ "$line_data" ]; then
 	    # Skip command #
@@ -134,69 +102,134 @@ function get_module_list
 	    ((++i))
 	fi
     done < $1
+ 
     echo ${entry[@]}
 }
 
 
 
-function setup_base_require
+function setup_base
+{
+    local git_status=$1; shift;
+    for rep in  ${base_list[@]}; do
+	if [ "${BOOL_GIT_CLONE}" = "$git_status" ]; then
+	    git_clone ${rep} ||  die 1 "${FUNCNAME[*]} : git clone ERROR at ${rep}: the target ${rep} may be exist, Please check it" ;
+	fi
+	if [[ $(checkIfDir "${rep}") -eq "$EXIST" ]]; then
+	    pushd ${rep}
+#	    checkout_e3_tags "target_path_test"
+	    make init   ||  die 1 "${FUNCNAME[*]} : MAKE init ERROR at ${rep}: Please check it" ;
+	    make env
+	    make pkgs
+	    make patch  ||  die 1 "${FUNCNAME[*]} : MAKE patch ERROR at ${rep}: Please check it" ;
+	    popd
+	else
+	    die 1 "${FUNCNAME[*]} : ${rep} doesn't exist";
+	fi
+	
+    done
+}
+
+
+
+function build_base
+{
+
+    sudo -v
+    for rep in  ${base_list[@]}; do
+	if [[ $(checkIfDir "${rep}") -eq "$EXIST" ]]; then
+	    pushd ${rep}
+#	    checkout_e3_tags "target_path_test"
+	    make build ||  die 1 "${FUNCNAME[*]} : Building Error at ${rep}: Please check the building error" ;
+	    popd
+	else
+	    die 1 "${FUNCNAME[*]} : ${rep} doesn't exist";
+	fi
+    done
+}
+
+
+function clean_base
+{
+
+    local rep;
+    for rep in  ${base_list[@]}; do
+	if [[ $(checkIfDir "${rep}") -eq "$EXIST" ]]; then
+	    echo "Cleaning .... $rep"
+	    sudo rm -rf ${SC_TOP}/${rep}
+	else
+	    printf " %20s: SKIP %20s, we cannot find it\n"  "${FUNCNAME[*]}"  "${rep}"
+	fi
+    done
+}
+
+
+
+
+
+function setup_require
 {
     local git_status=$1; shift;
     for rep in  ${require_list[@]}; do
 	if [ "${BOOL_GIT_CLONE}" = "$git_status" ]; then
-	    git_clone ${rep}
+	    git_clone ${rep}  ||  die 1 "${FUNCNAME[*]} : git clone ERROR at ${rep}: the target ${rep} may be exist, Please check it" ;
 	fi
-	pushd ${rep}
-	make init || die 1 "Init ERROR : Please check your ${rep}"  ;
-	make env
-	if [ "${rep}" = "e3-base" ]; then
-	    make pkgs || die 2 "PKGS ERROR : Please check your pkgs "  ;
+	if [[ $(checkIfDir "${rep}") -eq "$EXIST" ]]; then
+	    pushd ${rep}
+#	    checkout_e3_tags "target_path_test"
+	    make init ||  die 1 "${FUNCNAME[*]} : MAKE init ERROR at ${rep}: Please check it" ;
+	    make env
+	    popd
+	else
+	    die 1 "${rep} doesn't exist";
 	fi
-	popd
     done
 }
 
-function build_base_require
+
+
+function build_require
 {
 
     sudo -v
     
-    
     for rep in  ${require_list[@]}; do
-	pushd ${rep}
-	make build || die 3 "$BUILD ERROR : ${rep} "  ;
-	if [ "${rep}" = "e3-require" ]; then
-	    make install || die 4 "INSTALLL ERROR : Please check ${rep} "  ;
+	if [[ $(checkIfDir "${rep}") -eq "$EXIST" ]]; then
+	    pushd ${rep}
+#	    checkout_e3_tags "target_path_test"
+	    make build   ||  die 1 "${FUNCNAME[*]} : Building Error at ${rep}: Please check the building error" ;
+	    make install ||  die 1 "${FUNCNAME[*]} : MAKE INSTALL ERROR at ${rep}: Please check it" ;
+	    popd
+	else
+	    die 1 "${rep} doesn't exist";
 	fi
-	popd
     done
 }
 
 
-#
-# Mandatory step in order to compile modules
-# Since we have the global env variables,
-# it should be OK to use them
-#
-  
-function setup_env
+function clean_require
 {
-    local git_status=$1;  shift;
-    
-    if [ "${BOOL_GIT_CLONE}" = "$git_status" ]; then
-	git_clone ${env}
-    fi
-
-    source ${env}/setE3Env.bash
-
+    local rep;
+    for rep in  ${require_list[@]}; do
+	if [[ $(checkIfDir "${rep}") -eq "$EXIST" ]]; then
+	    echo "Cleaning .... $rep"
+	    make uninstall 
+	    sudo rm -rf ${SC_TOP}/${rep}
+	else
+	    printf " %20s: SKIP %20s, we cannot find it\n"  "${FUNCNAME[*]}"  "${rep}"
+	fi
+    done
 }
+
+
+
 
 
 
 function print_list
 {
-    local array=$1; shift;
-    for a_list in ${array[@]}; do
+    local a_list;
+    for a_list in ${module_list[@]}; do
 	printf " %s\n" "$a_list";
     done
 }
@@ -207,60 +240,37 @@ function setup_modules
     local rep;
     for rep in  ${module_list[@]}; do
 	if [ "${BOOL_GIT_CLONE}" = "$git_status" ]; then
-	    git_clone ${rep}
+	    git_clone ${rep}  ||  die 1 "${FUNCNAME[*]} : git clone ERROR at ${rep}: the target ${rep} may be exist, Please check it" ;
 	fi
-	pushd ${rep}
-	make init ||  die 1 "Init ERROR : Please check your ${rep}"  ;  
-	make env
-	popd
+	if [[ $(checkIfDir "${rep}") -eq "$EXIST" ]]; then
+	    pushd ${rep}
+#	    checkout_e3_tags "target_path_test"
+	    make init ||  die 1 "${FUNCNAME[*]} : MAKE init ERROR at ${rep}: Please check it" ; 
+	    make env
+	    make patch
+	    popd
+	else
+	    die 1 "${FUNCNAME[*]} : ${rep} doesn't exist";
+	fi
     done
 
 }
 
 function build_modules
 {
-    yes_or_no_to_go "SUDO is required"
 
-    sudo -v
     for rep in  ${module_list[@]}; do
-	pushd ${rep}
-	make build || die 3 "$BUILD ERROR : ${rep} "  ;
-	sudo make install  || die 4 "INSTALLL ERROR : Please check ${rep} "  ;
-	popd
+	if [[ $(checkIfDir "${rep}") -eq "$EXIST" ]]; then
+	    pushd ${rep}
+#	    checkout_e3_tags "target_path_test"
+	    make build    ||  die 1 "${FUNCNAME[*]} : Building Error at ${rep}: Please check the building error" ;
+	    make install  ||  die 1 "${FUNCNAME[*]} : MAKE INSTALL ERROR at ${rep}: Please check it" ;
+	    popd
+	else
+	    die 1 "${FUNCNAME[*]} : ${rep} doesn't exist";
+	fi
     done
 }
-
-function build_modules_without_question
-{
-
-    sudo -v
-    for rep in  ${module_list[@]}; do
-	pushd ${rep}
-	make build
-	sudo make install
-	popd
-    done
-}
-
-
-
-function clean_env
-{
-    echo "Cleaning .... e3-env"
-    sudo rm -rf ${SC_TOP}/e3-env
-}
-
-
-function clean_base_require
-{
-
-    local rep;
-    for rep in  ${require_list[@]}; do
-	echo "Cleaning .... $rep"
-	sudo rm -rf ${SC_TOP}/${rep}
-    done
-}
-
 
 
 function clean_modules
@@ -268,11 +278,15 @@ function clean_modules
     local rep;
     sudo -v;
     for rep in  ${module_list[@]}; do
-	echo "Cleaning .... $rep"
-	pushd ${rep}
-	make uninstall
-	popd
-	sudo rm -rf ${SC_TOP}/${rep}
+	if [[ $(checkIfDir "${rep}") -eq "$EXIST" ]]; then
+	    echo "Cleaning .... $rep"
+	    pushd ${rep}
+	    make uninstall 
+	    popd
+	    sudo rm -rf ${SC_TOP}/${rep}
+	else
+	    printf " %20s: SKIP %20s, we cannot find it\n"  "${FUNCNAME[*]}"  "${rep}"
+	fi
     done
 }
 
@@ -281,21 +295,29 @@ function clean_modules
 function git_pull
 {
     local rep;
+
+    for rep in  ${base_list[@]}; do
+	if [[ $(checkIfDir "${rep}") -eq "$EXIST" ]]; then
+	    pushd ${rep}
+	    git pull
+	    popd
+	fi
+    done
+    
     for rep in  ${require_list[@]}; do
-	pushd ${rep}
-	git pull
-	popd
+	if [[ $(checkIfDir "${rep}") -eq "$EXIST" ]]; then
+	    pushd ${rep}
+	    git pull
+	    popd
+	fi
     done
 
-    pushd ${env}
-    git pull
-    popd
-   
-    
     for rep in  ${module_list[@]}; do
-	pushd ${rep}
-	git pull
-	popd
+	if [[ $(checkIfDir "${rep}") -eq "$EXIST" ]]; then
+	    pushd ${rep}
+	    git pull
+	    popd
+	fi
     done
 }
    
@@ -306,9 +328,13 @@ function git_add
     local rep;
     local git_add_file=$1; shift;
     for rep in  ${module_list[@]}; do
-	pushd ${rep}
-	git add ${git_add_file}
-	popd
+	if [[ $(checkIfDir "${rep}") -eq "$EXIST" ]]; then
+	    pushd ${rep}
+	    git add ${git_add_file}
+	    popd
+	else
+	    printf " %20s: SKIP %20s, we cannot find it\n"  "${FUNCNAME[*]}"  "${rep}"
+	fi
     done
 }
    
@@ -318,9 +344,13 @@ function git_commit
     local rep;
     local git_commit_comment=$1; shift;
     for rep in  ${module_list[@]}; do
-	pushd ${rep}
-	git commit -m "${git_commit_comment}"
-	popd
+	if [[ $(checkIfDir "${rep}") -eq "$EXIST" ]]; then
+	    pushd ${rep}
+	    git commit -m "${git_commit_comment}"
+	    popd
+	else
+	    printf " %20s: SKIP %20s, we cannot find it\n"  "${FUNCNAME[*]}"  "${rep}"
+	fi
     done
 }
 
@@ -329,10 +359,16 @@ function git_push
 {
     local rep;
     for rep in  ${module_list[@]}; do
-	pushd ${rep}
-	git push
-	popd
+	if [[ $(checkIfDir "${rep}") -eq "$EXIST" ]]; then
+	    pushd ${rep}
+	    git push
+	    popd
+	else
+	    printf " %20s: SKIP %20s, we cannot find it\n"  "${FUNCNAME[*]}"  "${rep}"
+	fi  
     done
+    
+    
 }
 
 
@@ -348,13 +384,13 @@ function git_push
 
 function module_loading_test_on_iocsh
 {
-    source ${SC_TOP}/e3-env/setE3Env.bash
+    source ${SC_TOP}/e3-require/tools/setE3Env.bash
 
     local IOC_TEST=/tmp/module_loading_test.cmd
     
     {
 	local PREFIX_MODULE="EPICS_MODULE_NAME:="
-	local PREFIX_LIBVERSION="export LIBVERSION:="
+	local PREFIX_LIBVERSION="E3_MODULE_VERSION:="
 	local mod=""
 	local ver=""
 	printf "var requireDebug 1\n";
@@ -365,14 +401,27 @@ function module_loading_test_on_iocsh
 		elif [[ $line =~ "${PREFIX_MODULE}" ]] ; then
 		    mod=${line#$PREFIX_MODULE}
 		fi
-	    done < ${SC_TOP}/${rep}/configure/CONFIG
+	    done < ${SC_TOP}/${rep}/configure/CONFIG_MODULE
+	    if [ "${mod}" = "StreamDevice" ]; then
+		mod="stream"
+	    fi
 	    printf "#\n#\n"
 	    printf "# >>>>>\n";
 	    printf "# >>>>> MODULE Loading ........\n";
 	    printf "# >>>>> MODULE NAME ..... ${mod}\n";
 	    printf "# >>>>>        VER  ..... ${ver}\n";
 	    printf "# >>>>>\n";
-	    printf "require ${mod}, ${ver}\n";
+
+	    if [[ ${mod} == AD* ]]; then
+		printf "#--------------------------------------- \n";
+		printf "# In ADSupport, ADCore, and ADSimDector, \n";
+		printf "# Simply ignore the following errors : \n";
+		printf " epicsEnvSet(\"TOP\",\"${SC_TOP}/${rep}\")\n"
+		printf " cd ${SC_TOP}/${rep}\n"
+		printf " < ${SC_TOP}/${rep}/cmds/load_libs.cmd\n"
+		printf "#--------------------------------------- \n";
+	    fi
+	    printf "require ${mod},${ver}\n";
 	    printf "# >>>>>\n";
 	    printf "#\n#\n"
 	done
@@ -382,46 +431,218 @@ function module_loading_test_on_iocsh
     exec iocsh.bash ${IOC_TEST}
 }
 
+function base_all
+{
+    clean_base;
+    setup_base "TRUE";
+    build_base;
+}
 
-#env_reset
+function req_all
+{
+    clean_require;
+    setup_require "TRUE";
+    build_require;
+}
 
-module_list=$(get_module_list ${SC_TOP}/configure/MODULES)
+function mod_all
+{
+    clean_modules;
+    setup_modules  "TRUE";
+    build_modules;
+}
+
+function usage
+{
+    {
+	echo "";
+	echo "Usage    : $0 [ -g <group_name> ] <option> ";
+	echo "";
+	echo " < group_name > ";
+	echo ""
+	echo "           common : epics modules"
+	echo "           timing : mrf timing    related modules";
+	echo "           ifc    : ifc platform  related modules";
+	echo "           ecat   : ethercat      related modules";
+	echo "           area   : area detector related modules";
+	echo "           test   : common, timing, ifc modules";
+	echo "           test2  : common, timing, area modules";
+	echo "           jhlee  : common, timing, ifc, area modules";
+	echo "           all    : common, timing, ifc, ecat, area modules";
+	echo "";
+	echo " < option > ";
+	echo "";
+      
+	echo "           env    : Print enabled Modules";
+	echo ""
+	echo "           cbase  : Clean Base";
+	echo "           ibase  : Init  Base ";
+	echo "           bbase  : Build, Install Base";
+	echo "            base  : cbase, ibase, bbase";
+	echo ""           
+	echo ""
+	echo "           creq   : Clean Require";
+	echo "           ireq   : Init  Require";
+	echo "           breq   : Build, Install Require";
+	echo "            req   : creq, ireq, breq";
+	echo ""           
+	echo ""
+	echo "           cmod   : Clean Modules";
+	echo "           imod   : Init  Modules";
+	echo "           bmod   : Build, Install Modules";
+	echo "            mod   : cmod, imod, bmod";
+	echo ""
+	echo "           load   : Load all installed Modules into iocsh.bash";
+	echo ""           
+	echo ""    
+	
+	echo "  Examples : ";
+	echo ""    
+	echo "          $0 env";
+	echo "          $0 -g timing env";
+	echo "          $0 base";
+        echo "          $0 req";
+	echo "          $0 -g ecat cmod";
+	echo "          $0 -g timing load";
+	echo "   ";       
+	echo "";
+	
+    } 1>&2;
+    exit 1; 
+}
+
+ 
+while getopts " :g:" opt; do
+    case "${opt}" in
+	g)
+	    GROUP_NAME=${OPTARG}
+	    ;;
+	*)
+	    usage
+	    ;;
+    esac
+done
+shift $((OPTIND-1))
+
+case "${GROUP_NAME}" in
+    common)
+	module_list+=( "$(get_module_list ${SC_TOP}/configure/MODULES_COMMON)" )
+	;;
+    timing*)
+	module_list+=( "$(get_module_list ${SC_TOP}/configure/MODULES_TIMING)" )
+	;;
+    ifc)
+	module_list+=( "$(get_module_list ${SC_TOP}/configure/MODULES_IFC)" )
+	;;
+    ecat)
+	module_list+=( "$(get_module_list ${SC_TOP}/configure/MODULES_ECAT)" )
+	;;
+    area)
+	module_list+=( "$(get_module_list ${SC_TOP}/configure/MODULES_AD)" )
+	;;
+    test)
+	module_list+=( "$(get_module_list ${SC_TOP}/configure/MODULES_COMMON)" )
+	module_list+=( "$(get_module_list ${SC_TOP}/configure/MODULES_TIMING)" )
+	module_list+=( "$(get_module_list ${SC_TOP}/configure/MODULES_IFC)"    )
+#	module_list+=( "$(get_module_list ${SC_TOP}/configure/MODULES_AD)"     )
+	;;
+    test2)
+	module_list+=( "$(get_module_list ${SC_TOP}/configure/MODULES_COMMON)" )
+	module_list+=( "$(get_module_list ${SC_TOP}/configure/MODULES_TIMING)" )
+#	module_list+=( "$(get_module_list ${SC_TOP}/configure/MODULES_IFC)"    )
+	module_list+=( "$(get_module_list ${SC_TOP}/configure/MODULES_AD)"     )
+	;;
+    jhlee)
+	module_list+=( "$(get_module_list ${SC_TOP}/configure/MODULES_COMMON)" )
+	module_list+=( "$(get_module_list ${SC_TOP}/configure/MODULES_TIMING)" )
+	module_list+=( "$(get_module_list ${SC_TOP}/configure/MODULES_IFC)"    )
+	module_list+=( "$(get_module_list ${SC_TOP}/configure/MODULES_AD)"     )
+	echo ""
+	;;
+    all)
+	module_list+=( "$(get_module_list ${SC_TOP}/configure/MODULES_COMMON)" )
+	module_list+=( "$(get_module_list ${SC_TOP}/configure/MODULES_TIMING)" )
+	module_list+=( "$(get_module_list ${SC_TOP}/configure/MODULES_IFC)"    )
+	module_list+=( "$(get_module_list ${SC_TOP}/configure/MODULES_ECAT)"   )
+	module_list+=( "$(get_module_list ${SC_TOP}/configure/MODULES_AD)"     )
+	;;
+    * )
+	module_list+=( "e3-iocStats" )
+	;;
+esac
+
+
+case "$1" in
+    *base)
+	echo ""
+	;;
+    *req)
+	echo ""
+	;;
+    clean)
+	echo ""
+	;;
+    *) 
+	echo ">> Selected Modules are :"
+	echo ${module_list[@]}
+	echo ""
+	;;
+
+esac
 
 
 case "$1" in
     all)
-	setup_base_require "TRUE"
-	build_base_require
-	setup_env          "TRUE"
-	setup_modules      "TRUE"
-	build_modules
-	setup_env
-	install_db
+	base_all;
+	req_all;
+	mod_all;
 	;;
-    travis)
-	setup_base_require "TRUE"
-	build_base_require
-	setup_env          "TRUE"
-	setup_modules      "TRUE"
-	build_modules_without_question
-	setup_env
-	install_db
-	;;
-    base)
-    	setup_base_require "TRUE"
-	build_base_require
-	;;
-    modules)
-	setup_env     "TRUE"
-	setup_modules "TRUE"
-	build_modules
+    clean)
+	clean_base;
+	clean_require;
+	clean_modules;
 	;;
     env)
-	print_list "${module_list[@]}"
+	echo ">> Vertical display for the selected modules :"
+	echo ""
+	print_list
+	echo ""
+	;;
+    cbase)
+	clean_base;
+	;;
+    ibase)
+	setup_base "TRUE";
+	;;
+    bbase)
+	build_base;
+	;;
+    base)
+	base_all;
+	;;
+    creq)
+	clean_require
+	;;
+    ireq)
+	setup_require "TRUE";
+	;;
+    breq)
+	build_require;
+	;;
+    req)
+	req_all;
+	;;
+    cmod)
+	clean_modules
+	;;
+    imod)
+	setup_modules  "TRUE"
 	;;
     bmod)
-	setup_env  
 	build_modules
+	;;
+    mod)
+	mod_all
 	;;
     pull)
 	git_pull
@@ -435,29 +656,14 @@ case "$1" in
     push)
 	git_push
 	;;
-    rmod)
-	clean_modules
-	setup_env      "TRUE"
-	setup_modules  "TRUE"
-	build_modules
-	;;
-    clean)
-	clean_base_require
-	clean_env
-	clean_modules
-	;;
-    cmod)
-	clean_modules
-	;;
     db)
-	setup_env
 	install_db
 	;;
     load)
 	module_loading_test_on_iocsh
 	;;
     *)
-	help
+	usage
 	;;
 esac
